+++ elasticjob-infra/elasticjob-registry-center/src/test/java/org/apache/shardingsphere/elasticjob/reg/zookeeper/ZookeeperConfigurationTest.java
@@ -29,6 +29,7 @@ public final class ZookeeperConfigurationTest {
         ZookeeperConfiguration zkConfig = new ZookeeperConfiguration("localhost:2181", "myNamespace");
         assertThat(zkConfig.getServerLists(), is("localhost:2181"));
         assertThat(zkConfig.getNamespace(), is("myNamespace"));
+        assertThat(zkConfig.isEnsembleTracker(), is(true));
         assertThat(zkConfig.getBaseSleepTimeMilliseconds(), is(1000));
         assertThat(zkConfig.getMaxSleepTimeMilliseconds(), is(3000));
         assertThat(zkConfig.getMaxRetries(), is(3));